heliora_backend/pyproject.toml
[project]
name = "heliora-backend"
version = "0.1.0"
description = "Heliora backend scaffold"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
  "fastapi>=0.115,<1.0",
  "uvicorn[standard]>=0.30,<1.0",
  "pydantic-settings>=2.4,<3.0",
  "pika>=1.3,<2.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["app", "tests"]

[tool.ruff.lint]
select = ["E", "F"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
files = ["app"]
ignore_missing_imports = true
check_untyped_defs = true
no_implicit_optional = true
warn_unused_ignores = true
warn_redundant_casts = true
pretty = true

[tool.coverage.run]
source = ["app"]
branch = true
omit = ["app/__init__.py"]

[tool.coverage.report]
fail_under = 70
show_missing = true
skip_covered = true
